Correct type

+++ src/lib/P11Attributes.cpp
@@ -1637,7 +1637,7 @@ CK_RV P11AttrModulus::updateAttr(Token *token, bool isPrivate, CK_VOID_PTR pValu
 	OSAttribute* attrClass = osobject->getAttribute(CKA_CLASS);
 	if (op == OBJECT_OP_CREATE && attrClass != NULL && attrClass->getUnsignedLongValue() == CKO_PUBLIC_KEY)
 	{
-		osobject->setAttribute(CKA_MODULUS_BITS, plaintext.bits());
+		osobject->setAttribute(CKA_MODULUS_BITS, (unsigned long)plaintext.bits());
 	}
 
 	return CKR_OK;
